[STRUCTURAL] Extract Class: skill runtime construction into internal/skillruntime
Phase 3 slice 4, and the largest so far. createSkillRuntime built the
loader, the integration objects, five backend adapters, the backend and
sandbox factories, the runtime itself, and then ran discovery and trigger
activation — about 150 lines in main.go, plus the adapter types and the
noop prompt backend they depend on. Which backends exist, what each is
handed, and which skills switch on are product behaviour.

Composition inputs become an Options struct the caller fills: config,
mode, working directory, config directory, the skill selection the flags
resolved to, and whether those are pre-approved. main.go keeps a small
createSkillRuntime that translates --skills and --approve-skills into
that struct.

Two things this removes rather than moves:

- createSkillRuntime resolved the config directory itself with
  os.UserHomeDir, duplicating configDir(), even though all three callers
  had already resolved it. It is now a parameter.
- registerBuiltinSkillPrompts had two more callers in skill.go, which
  built their own loaders. It is exported as RegisterBuiltinPrompts so
  the CLI subcommands and the runtime register the same built-ins.

main.go: 3,069 -> 2,795 lines, and drops nine imports — the builtin skill
packages, all four backends, sandbox, and skillsdk.

The adapter tests move in-package: they exercise unexported wiring, not
this package's contract.
