+++ cmd/rubichan/shell.go
@@ -114,7 +114,7 @@ func runShell() error {
 	// Create skill runtime.
 	ctx := context.Background()
 	opts = appendKnowledgeGraphOption(ctx, opts, cwd)
-	rt, storeCloser, err := createSkillRuntime(ctx, registry, p, cfg, "shell", cwd)
+	rt, storeCloser, err := createSkillRuntime(ctx, registry, p, cfg, "shell", cwd, cfgDir)
 	if err != nil {
 		return fmt.Errorf("creating skill runtime: %w", err)
 	}
